tools/seo-automation/requirements.txt
# Python 3.6+ (older hosts) and 3.11+ (GitHub Actions / local venv).
google-api-python-client==2.52.0
google-auth==2.14.1
